급여 상세 내역 페이지 컴포넌트 분리 급여 상세내역 카드 컴포넌트 분리 월별 상세 확인 버튼 컴포넌트 분리 styled.tsx 확장자 ts로 변경 Github issue #54
