+++ src/pages/salaryDetail/MoveMonth.style.ts
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const Movemonth = styled.div`
+  height:30px;
+  margin-top:1.3rem;
+  text-align:center;
+
+  .css-i4bv87-MuiSvgIcon-root{
+      height:0.8em;
+  }
+
+  .up-date{
+    font-weight:var(--font-weight-bold);
+    font-size:1.3rem;
+    margin:0 5px;
+
+    .css-i4bv87-MuiSvgIcon-root{
+      height:0.7em;
+    }
+  }
+`;
